% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleandata.R
\name{cleandata}
\alias{cleandata}
\title{Clean up your dataset}
\usage{
cleandata(x = cutoffvalue:::exampledata)
}
\arguments{
\item{x}{Your dataset specified as "DatasetName$ColumnName"; data should be a single column of log-transformed data with a column header.}
}
\value{
Returns your cleaned dataset along with the minimum and maximum values from that dataset.
}
\description{
This function cleans up your dataset after it has been added to the Global Environment. It will remove any blank rows and determine the minimum and maximum values of the dataset.
}
\examples{
mydata <- cleandata(cutoffvalue:::exampledata)

}
