% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datamodel.R
\name{datamodel}
\alias{datamodel}
\title{Generate Normal Mix EM Model for Data}
\usage{
datamodel(mydata = "cutoffvalue/extdata/exampledata.xlsx")
}
\arguments{
\item{mydata}{Your data file; data should be organized in as a single column of log- or natural log-transformed data without a
column header}
}
\value{
Model parameters and lower index
}
\description{
Generate Normal Mix EM Model for Data
}
\examples{
Mcmodel <- datamodel()
}
