% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findCutoff.R
\name{find.cutoff}
\alias{find.cutoff}
\title{Find Cutoff Value}
\usage{
find.cutoff(proba = 0.5, i = index.lower)
}
\arguments{
\item{proba}{The probability value you want for the cutoff value to represent}

\item{i}{Determines the location, i.e., index of the (first) minimum of the model$mu}
}
\value{
the cutoff value for the indicated probability value
}
\description{
Find Cutoff Value
}
\examples{
cutoff <- c(find.cutoff())
mydata.lower <- -2
mydata.upper <- 2
}
