% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findCutoff.R
\name{findCutoff}
\alias{findCutoff}
\title{Find Cutoff Value}
\usage{
findCutoff(mydata = "data/exampledata.xlsx", proba = 0.5, i = model$indexLower)
}
\arguments{
\item{proba}{The probability value you want for the cutoff value to represent}

\item{i}{Determines the location, i.e., index of the (first) minimum of the model$mydata$mu}
}
\value{
the cutoff value for the indicated probability value
}
\description{
Find Cutoff Value
}
\examples{
cutoff <- find.cutoff()
}
