% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{fitParams}
\alias{fitParams}
\title{Histogram for fitting curves}
\usage{
fitParams(mydata = "data/exampledata.xlsx")
}
\arguments{
\item{mydata}{Your data file; data should be organized in as a single column of log- or natural log-transformed data without a
column header}
}
\value{
Basic histogram of dataset along with various parameters used to fit curves
}
\description{
Histogram for fitting curves
}
\examples{
fit <- fit()
}
