% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importData.R
\name{import.data}
\alias{import.data}
\title{Import Data}
\usage{
import.data(x)
}
\arguments{
\item{x}{An Excel spreadsheet with headerless, log-transformed data in the first column}
}
\value{
A list of values from the Excel spreadsheet that has blanks removed
}
\description{
Import Data
}
\examples{
rawdata <- "exampledata.xlsx"
import.data(rawdata)
}
