% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importData.R
\name{importdata}
\alias{importdata}
\title{This function imports your data set.}
\usage{
importdata(mydata)
}
\arguments{
\item{mydata}{Your data file; data should be organized in as a single column of log- or natural log-transformed data without a
column header}
}
\value{
Returns a list of values from the raw data in your Excel spreadsheet that has blanks removed.
}
\description{
This function imports your data set.
}
\examples{
mydata <- importdata()
}
