% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modes.R
\name{modes}
\alias{modes}
\title{Test Data Modality}
\usage{
modes(mydata = "data/exampledata.xlsx")
}
\arguments{
\item{x}{The data}
}
\value{
The p-value for whether or not the dataset is unimodal
}
\description{
Test Data Modality
}
\examples{
mode.pvalue <- modes()
}
