% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mydata.R
\docType{data}
\name{mydata}
\alias{mydata}
\title{Log 11-keto testosterone values from age 2 male spring chinook}
\format{
\subsection{\code{mydata}}{

A list of 97 values:
\describe{
\item{log 11-KT}
...
}
}
}
\source{
\url{https://doi.org/10.1002/tafs.10097}
}
\usage{
mydata
}
\description{
A subset of data from a study investigating precocious maturation in hatchery-origin, age 2 male spring chinook in Cle Elum, WA
}
\keyword{datasets}
