% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{myDataPlot}
\alias{myDataPlot}
\title{Plot Histogram with Fit Lines}
\usage{
myDataPlot(
  maintitle = "",
  xlabel = "Plasma [11-KT] (ng/mL)",
  cutofflabel = "Minijack cutoff",
  cutoffunits = "ng/mL"
)
}
\arguments{
\item{maintitle}{Title for the graph, default is blank}

\item{xlabel}{Label for the x-axis, default is "Plasma \link{11-KT} (ng/mL)"}

\item{cutofflabel}{Label for the cutoff value, default is "Minijack cutoff"}

\item{cutoffunits}{Label for the units, default is "ng/mL"}
}
\value{
Histogram with colored lines depicting upper and lower modes along with a cutoff line to delineate between the two modes
}
\description{
Plot Histogram with Fit Lines
}
\examples{
plot()
plot("Example Graph Title", "Example X-Axis", "Cutoff Label", "UNITS")
}
