% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rawdata.R
\docType{data}
\name{rawdata}
\alias{rawdata}
\title{Juvenile male spring chinook plasma 11-KT levels}
\format{
\subsection{\code{rawdata}}{

A data frame with 98 rows and 1 column:
\describe{
\item{11-KT Values}
}
}
}
\source{
\url{https://doi.org/10.1002/tafs.10097}
}
\usage{
rawdata
}
\description{
A subset of plasma 11-KT values from age-2 juvenile male spring chinook collected to evaluate precocious maturation
}
\keyword{datasets}
