% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curves.R
\name{curves}
\alias{curves}
\title{This function fits curves to the data set. As a matter of course, this function runs the importData, datamodel, and fit functions for you, negating the need to run those separately.}
\usage{
curves(mydata)
}
\arguments{
\item{mydata}{Your data file; data should be organized in as a single column of log- or natural log-transformed data without a
column header}
}
\value{
Returns three sets of values, which are used to plot curves for the lower and upper modes.
}
\description{
This function fits curves to the data set. As a matter of course, this function runs the importData, datamodel, and fit functions for you, negating the need to run those separately.
}
\examples{
curves <- curves()
}
