% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curves.R
\name{curves}
\alias{curves}
\title{Fit curves to each mode of the dataset}
\usage{
curves(x = cutoffvalue:::exampledata)
}
\arguments{
\item{x}{Your dataset specified as "DatasetName$ColumnName"; data should be a single column of log-transformed data with a column header.}
}
\value{
Returns three sets of values, which are used to plot curves for the lower and upper modes.
}
\description{
This function fits curves to the dataset. As a matter of course, this function runs the cleanData, datamodel, and fit functions for you, negating the need to run those separately.
}
\examples{
curves <- curves(cutoffvalue:::exampledata)
}
