% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curves.R
\name{curves}
\alias{curves}
\title{Fit curves to dataset}
\usage{
curves(mydata = "cutoffvalue/extdata/exampledata.xlsx")
}
\arguments{
\item{mydata}{Your data file; data should be organized in as a single column of log- or natural log-transformed data without a
column header}
}
\value{
Three sets of data used to plot curves for the lower and upper modes
}
\description{
Fit curves to dataset
}
\examples{
curves <- curves()
}
