% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curves.R
\name{curves}
\alias{curves}
\title{Fit curves to dataset}
\usage{
curves(x)
}
\arguments{
\item{x}{Dataset}
}
\value{
Three sets of data used to plot curves for the lower and upper modes
}
\description{
Fit curves to dataset
}
\examples{
curves <- curves()
}
