% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curves.R
\name{curves}
\alias{curves}
\title{Fit curves to each mode of the dataset}
\usage{
curves(x)
}
\arguments{
\item{rawdata}{Your data file; data should be organized in an Excel document as a single column of log- or natural log-transformed data without a
column header.}
}
\value{
Returns three sets of values, which are used to plot curves for the lower and upper modes.
}
\description{
This function fits curves to the dataset. As a matter of course, this function runs the cleanData, datamodel, and fit functions for you, negating the need to run those separately.
}
\examples{
curves <- curves(rawdata)
}
