% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datamodel.R
\name{dataModel}
\alias{dataModel}
\title{Generate Normal Mix EM Model for Data}
\usage{
dataModel(mydata = "data/exampledata.xlsx")
}
\arguments{
\item{x}{Dataset}
}
\value{
Model parameters and lower index
}
\description{
Generate Normal Mix EM Model for Data
}
\examples{
model <- datamodel()
}
