% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findCutoff.R
\name{findcutoff}
\alias{findcutoff}
\title{This function finds the cutoff value between the modes of bimodally distributed data at the indicated probability value. As a matter of course, this function runs the importData and datamodel functions for you, negating the need to run those separately.}
\usage{
findcutoff(mydata = "R/exampledata.xlsx", proba = 0.5, i = model$indexLower)
}
\arguments{
\item{mydata}{Your data file; data should be organized in as a single column of log- or natural log-transformed data without a
column header}

\item{proba}{The probability value you want the cutoff value to represent}

\item{i}{Determines the location, i.e., index of the (first) minimum of the model$mydata$mu}
}
\value{
Returns the cutoff value for the indicated probability value.
}
\description{
This function finds the cutoff value between the modes of bimodally distributed data at the indicated probability value. As a matter of course, this function runs the importData and datamodel functions for you, negating the need to run those separately.
}
\examples{
cutoff <- findcutoff()
}
