% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{fitParams}
\alias{fitParams}
\title{Histogram for fitting curves}
\usage{
fitParams(x)
}
\arguments{
\item{x}{Dataset}
}
\value{
Basic histogram of dataset along with various parameters used to fit curves
}
\description{
Histogram for fitting curves
}
\examples{
fit <- fit()
}
