% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{fitparams}
\alias{fitparams}
\title{Generate a histogram, with associated parameters, from the dataset}
\usage{
fitparams(x = cutoffvalue:::exampledata, breaks = 15)
}
\arguments{
\item{x}{Your dataset specified as "DatasetName$ColumnName"; data should be a single column of log-transformed data with a column header.}
}
\value{
Returns a basic histogram for the data set along with various parameters used to generate curves for the final plot.
}
\description{
This function produces a histogram of the data that will be used for fitting curves. As a matter of course, this function runs the cleanData and datamodel functions for you, negating the need to run those separately.
}
\examples{
fit <- fitparams(cutoffvalue:::exampledata)
}
