% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importData.R
\name{importData}
\alias{importData}
\title{Import Data}
\usage{
importData(mydata = "inst/extdata/exampledata.xlsx")
}
\arguments{
\item{mydata}{Your data file; data should be organized in as a single column of log- or natural log-transformed data without a
column header}
}
\value{
A list of values from the Excel spreadsheet that has blanks removed
}
\description{
Import Data
}
\examples{
mydata <- importData()
}
