% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importData.R
\name{importData}
\alias{importData}
\title{Import Data}
\usage{
importData(mydata = "data/exampledata.xlsx")
}
\arguments{
\item{mydata}{}
}
\value{
A list of values from the Excel spreadsheet that has blanks removed
}
\description{
Import Data
}
\examples{
import.data(rawdata)
}
