% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importData.R
\name{importdata}
\alias{importdata}
\title{Import Data}
\usage{
importdata(mydata)
}
\arguments{
\item{mydata}{Your data file; data should be organized in as a single column of log- or natural log-transformed data without a
column header}
}
\value{
A list of values from the Excel spreadsheet that has blanks removed
}
\description{
Import Data
}
\examples{
mydata <- importdata()
}
