% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modes.R
\name{modes}
\alias{modes}
\title{Test Data Modality}
\usage{
modes(mydata = "inst/extdata/exampledata.xlsx")
}
\arguments{
\item{mydata}{Your data file; data should be organized in as a single column of log- or natural log-transformed data without a
column header}
}
\value{
The p-value for whether or not the dataset is unimodal
}
\description{
Test Data Modality
}
\examples{
modeTest <- modes()
}
