% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modes.R
\name{modes}
\alias{modes}
\title{Determines the modality of the dataset}
\usage{
modes(x = cutoffvalue:::exampledata)
}
\arguments{
\item{x}{Your dataset specified as "DatasetName$ColumnName"; data should be a single column of log-transformed data with a column header.}
}
\value{
Returns the p-value and excess mass statistic to determine whether the dataset is unimodal.
}
\description{
This function tests the dataset's modality. It uses the modetest function from the multimode package written by Ameijeiras-Alonso et al. (2019) to determine the excess mass test statistic and test the number of modes. As a matter of course, this function runs the cleanData function, negating the need to run it separately.
}
\examples{
modetest <- modes(cutoffvalue:::exampledata)
}
